Initial CLI

+++ src/loader/dir.rs
@@ -13,16 +13,27 @@ use super::BundleLoader;
 /// `BundleType::suffix()`
 ///
 /// If the bundles are loaded by ID, then the bundle name is assumed to be the ID with the corresponding extension
-/// i.e. `<ID>.bundle`, `<ID>.bin`, or `<ID>`
+/// i.e. `<ID>.bundle`, `<ID>.bin`, or `<ID>`. Otherwise, each file in the directory is treated as a bundle as long as
+/// it has an extension matching one of the ones returned by `BundleType::suffix()`, and the loader just loads (and removes)
+/// a random file from the directory
 #[derive(Debug, Clone)]
 pub struct DirLoader {
     path: PathBuf,
+    delete_after_load: bool,
 }
 
 impl DirLoader {
-    /// Creates a new `DirLoader` with the given path.
-    pub const fn new(path: PathBuf) -> Self {
-        Self { path }
+    /// Creates a new `DirLoader`
+    ///
+    /// Arguments
+    /// - `path`: The path to the directory to load the bundles from
+    /// - `delete_after_load`: A flag indicating whether the loaded bundle should be deleted from the directory after loading
+    ///   Only used when a bundle is loaded without a supplied ID (i.e. a random bundle)
+    pub const fn new(path: PathBuf, delete_after_load: bool) -> Self {
+        Self {
+            path,
+            delete_after_load,
+        }
     }
 }
 
@@ -68,7 +79,7 @@ impl BundleLoader for DirLoader {
 
             io::copy(&mut file, &mut write).context("Copying the bundle failed")?;
 
-            if id.is_none() {
+            if id.is_none() && self.delete_after_load {
                 fs::remove_file(&path).context("Removing the bundle failed")?;
             }
